requirements.txt
flask
flask_sqlalchemy
jinja2
flask_migrate
flask_bcrypt
flask_wtf
dotenv
selenium